fix(import): make JSONImportPluginTests link by extracting pure parsing into a shared file (#1536)
